+++ src/routes/Login.tsx
@@ -3,7 +3,8 @@ import { useForm } from 'react-hook-form';
 import { useHistory } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
 import styled from 'styled-components';
-import { fetchLogin, fetchUserInfo } from '../api';
+import { fetchUserInfo } from '../api/common';
+import { fetchLogin } from '../api/login';
 import { userInfoState } from '../atoms';
 import { LoginFormData } from '../types/login';
