+++ src/routes/board/NormalBoard.tsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { fetchPostList } from '../../api';
 import { userInfoState } from '../../atoms';
 import BoardHeader from '../../components/board/BoardHeader';
 import PostList from '../../components/board/PostList';
 import BoardListFooter from '../../components/board/BoardListFooter';
 import NavContainer from '../../components/board/NavContainer';
 import { BoardURLParams, Post } from '../../types/board';
+import { fetchPostList } from '../../api/board';
 
 function NormalBoard() {
   const [postList, setPostList] = useState<Post[]>([]);
